Fixed bugs found on initial unveiling in beers and the handler code.
Beers now counts correctly and does not die when it can't find a user. The code
for nonexistant users was bogus and the code for setting beers was not counting
dates.

Handlers were not breaking after finding users (linear scan, ack). This should
probably be switched to a map[string]user type.