Refs #51: Set up a test service
This service is put in place to run unit tests and possibly in the
future could be used as a console interface to the bot for other apps.

This is a sketch of what services might look like in the future, but is
only one idea. Services ought to have an interface that goes something
like recieving channel and sending channel in the end, but for now a
function will do.